- Hoe Auto replants selected crops from config.
Tools and crops are both configurable from the config file.

Once crops are fully grown using a hoe on the crop will break and replant the crop.

Crops cannot be broken by the hoe until they reach max growth. Durability is not taken off of the hoe when used on crops which cannot be broken.